+++ etna/transforms/encoders/mean_segment_encoder.py
@@ -1,24 +1,34 @@
-import reprlib
-from typing import Dict
 from typing import List
-from typing import Optional
 
 import numpy as np
 import pandas as pd
 
 from etna.transforms import IrreversibleTransform
-from etna.transforms.math.statistics import MeanTransform
+from etna.transforms.encoders.mean_encoder import MeanEncoderTransform
 
 
 class MeanSegmentEncoderTransform(IrreversibleTransform):
     """Makes expanding mean target encoding of the segment. Creates column 'segment_mean'."""
 
-    idx = pd.IndexSlice
+    _segment_column = "segment_column"
+    out_column = "segment_mean"
 
     def __init__(self):
         super().__init__(required_features=["target"])
-        self.mean_encoder = MeanTransform(in_column="target", window=-1, out_column="segment_mean")
-        self.global_means: Optional[Dict[str, float]] = None
+        self._mean_encoder = MeanEncoderTransform(
+            in_column=self._segment_column, mode="per-segment", out_column=self.out_column, smoothing=0
+        )
+
+    def _add_segment_column(self, df):
+        segments = df.columns.get_level_values("segment").unique()
+        flatten_segments = np.repeat(segments.values[np.newaxis, :], len(df), axis=0)
+        segment_values = pd.DataFrame(
+            data=flatten_segments,
+            columns=pd.MultiIndex.from_product([segments, [self._segment_column]]),
+            index=df.index,
+        )
+        df = pd.concat([df, segment_values], axis=1).sort_index(axis=1)
+        return df
 
     def _fit(self, df: pd.DataFrame) -> "MeanSegmentEncoderTransform":
         """
@@ -34,10 +44,8 @@ class MeanSegmentEncoderTransform(IrreversibleTransform):
         :
             Fitted transform
         """
-        self.mean_encoder._fit(df)
-        mean_values = df.loc[:, self.idx[:, "target"]].mean().to_dict()
-        mean_values = {key[0]: value for key, value in mean_values.items()}
-        self.global_means = mean_values
+        df = self._add_segment_column(df)
+        self._mean_encoder._fit(df)
         return self
 
     def _transform(self, df: pd.DataFrame) -> pd.DataFrame:
@@ -61,25 +69,11 @@ class MeanSegmentEncoderTransform(IrreversibleTransform):
         NotImplementedError:
             If there are segments that weren't present during training.
         """
-        if self.global_means is None:
-            raise ValueError("The transform isn't fitted!")
-
-        segments = df.columns.get_level_values("segment").unique().tolist()
-        new_segments = set(segments) - self.global_means.keys()
-        if len(new_segments) > 0:
-            raise NotImplementedError(
-                f"This transform can't process segments that weren't present on train data: {reprlib.repr(new_segments)}"
-            )
-
-        df = self.mean_encoder._transform(df)
-        segment = segments[0]
-        nan_timestamps = df[df.loc[:, self.idx[segment, "target"]].isna()].index
-        values_to_set = np.array([self.global_means[x] for x in segments])
-        # repetition isn't necessary for pandas >= 1.2
-        values_to_set = np.repeat(values_to_set[np.newaxis, :], len(nan_timestamps), axis=0)
-        df.loc[nan_timestamps, self.idx[:, "segment_mean"]] = values_to_set
-        return df
+        df = self._add_segment_column(df)
+        df_transformed = self._mean_encoder._transform(df)
+        df_transformed = df_transformed.drop(columns=[self._segment_column], level="feature")
+        return df_transformed
 
     def get_regressors_info(self) -> List[str]:
         """Return the list with regressors created by the transform."""
-        return ["segment_mean"]
+        return [self.out_column]